+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,16 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-01-XX
 
+## [0.34.0] - 2025-01-17
+
 ### Added
 
   - [ET-1417] `Draft::discard` - Discards a draft from the composer.
   - [ET-1685] `scroller_search` & `SearchScroller` to make server searches
   - More logging for the html transformations.
+  - [ET-679] `draft_undo_send` - Cancel sending of a message.
 
 ### Fixed
 
   - Drafts are moved to outbox before being sent.
   - [ET-503] Drafts can not be updated after being sent.
+
+### Changed
+
+  - `DraftSendStatus` - now includes whether it can be cancelled or not.
 
 ## [0.33.0] - 2025-01-17